Refactor author logic code
