+++ authors.json
@@ -1,1 +1,1 @@
-[{"id":814,"first_name":"Dolph","last_name":"Landergrin"},{"id":112,"first_name":"Hope","last_name":"Chirinda"}]
+[{"id":73,"first_name":"Dolph","last_name":"Landergreen"},{"id":344,"first_name":"Amen","last_name":"Musingarimi"}]
